# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "info@wirelessranking.com"
        },
        "version": "1.0.0"
    },
    "host": "ranking.dev.sapient.pro",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/countries": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "Get list of countries",
                "description": "Get list of countries",
                "operationId": "countries",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
        "/configs": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "Get config data for tests etc",
                "description": "Get config data for tests etc",
                "operationId": "configs",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
        "/contact": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get content for the contact page",
                "description": "Get content for the contact page",
                "operationId": "contact",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Send contact message",
                "description": "Send contact message",
                "operationId": "contact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Operator id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        }
    },
...
    "securityDefinitions": {
        "passport": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/oauth/token",
            "scopes": {}
        }
    }
}